refactor and implementation of client multi processing
